fix: which func SourceFile is using (new() => from_string). All tests passed.
