+++ src/tests/mod.rs
@@ -1,4 +1,5 @@
 mod compiler_test;
 mod lexical_test;
+mod literals_test;
 mod source_test;
 mod token_test;
